Altering merge_bams.sh to work correctly

+++ merge_bams.sh
@@ -17,12 +17,16 @@ SPEC=$1
 RAW=$(ls $SPEC*.sorted.bam)
 CLEAN=$(ls $SPEC*.dedup.bam)
 
-samtools merge $SPEC.raw.bam $RAW 
-if [ $? -eq 0 ]
+if [ ! -s $SPEC.raw.bam ]
 then
-	rm $SPEC*.sorted.bam
+	samtools merge $SPEC.raw.bam $RAW 
+	if [ $? -eq 0 ]
+	then
+		rm $SPEC*.sorted.bam
+	fi
 fi
-samtools merge -u - $CLEAN | samtools view -F 0x4 -f 0x2 -b > $SPEC.tmp.bam
+
+samtools merge -u - $CLEAN | samtools view -F 0x4 -f 0x2 -b - > $SPEC.tmp.bam
 java -Xmx2g -jar ~/sw/progs/picard-tools-1.121/MarkDuplicates.jar TMP_DIR=/scratch INPUT=$SPEC.tmp.bam OUTPUT=${SPEC}.clean.bam METRICS_FILE=${SPEC}.clean.metrics.txt REMOVE_DUPLICATES=true
 if [ $? -eq 0 ]
 then
